requirements.txt
PyPDF2
Spire.PDF
PyMuPDF
python-dotenv
openai
chromadb
google-api-python-client
certifi
beautifulsoup4
lxml
base64
webdriver-manager
pandas
openpyxl
aiohttp
pymongo
semantic-chunkers
tqdm